Add MethodResolution class
